Set sent back to 2
By setting the sent value of all existing campaign responses to 1, it
excludes them from the charts on the dashboard.  A value of 1 normally
indicates that the message did not go through, therefore, it does not
factor into stats regarding responses.  I think it makes most sense to
treat previous campaign emails as if they were successfully sent.

I also fixed a percentage rounding issue in the charts.