Improved Form entry caching and error message
Inline with issue #77 I've made it so that (where it makes sense) to
retain entered values from forms in the user module after an error
message is thrown.  I've also moved the error message into a new row in
the form highlighted in red instead of the pop-over.  Much more useable
and will be the template for which the other modules will follow.