Upgrade Fixes
Fixing a bad image link, added a new image that wasn't in the upgrade
package and setting the Editor module to a non-core module since it is
built-in now.  I think we should make it clear to the upgraded that
they can just delete the editor module in favor of the new one.