Reduced Goo.gl shortener code and fixed some misplaced files
The goo.gl shortener code was a script written by someone else, so to
avoid adding another license, I wrote a function to replace it that is
much simpler.  I uploaded all the necessary files in the upgrade
package and updated the install file as well.  I also fixed a broken
image link in the upgrade install file.  And a put the forgot password
file in the right location in the upgrade package…it was in the wrong
spot.