Issue #134 -> Display System Time & Allow for Time Zone Update
Now you can view the system time in the settings and a tooltip provides
a tip on how to change the system time.  However, it is not feasible to
change the system time within the app because this typically takes root
access.  However, changing the timezone is feasible.  So now the
timezone is stored in the settings table and this timezone setting
needs to be taken into account wherever time is utilized.