Downgrade jquery & Edit Campaign Work
deteremined that the highcharts graphs were not displaying correctly
due to a bug with newer versions of jquery…added an older version to
includes and addressed it instead of the newer jquery only on the
dashboard. Also did some work on the edit campaign functionality that
is introduced with the ability to schedule campaigns.