Upgrade from Settings Now!
You can now upload a zip file in the settings page and it will be
uploaded, the root of the zip file will be extracted into the root of
spt, permissions on the entire install will be cleaned up and the
temporary zip file deleted.  If an install.php script exists in the
root of the site, it will pass control to it.