Java and Flash Detection work
Can now detect java version and report it in the response pop-over.
Consolidated all software related versions into one hover over a new
icon.  All versions are exportable with the CSV file.  Flash work is
not complete…but the framework is there to support storing it once the
detection methodology is figured out.