Maybe fixed issue #160
this needs to be tested on an SSL install…but I think i've fixed it by
forcing cron jobs to point to http instead of https.  this should not
pose a risk since this traffic will never leave the localhost, but may
cause problems if someone is trying to force https.  I doubt anyone
would do this as you'd want to allow http for certain portions of spt.