Install script & SMTP edit script fixes
the install script was still looking into a now non-existent modules
directory for part of the SQL install...changed this to the new
"settings" directory.  If you made an edit to an SMTP server that had a
password…it would wipe the password out unless you entered it again.
Now it looks to see if it is empty or not.